pyproject.toml
[project]
name = "generargrafopetersen"
version = "0.1.0"
description = "Unified orchestrator for FlowToGraph and TermToGraph pipelines"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=6.1.0",
    "pandas>=3.0.1",
    "paramiko>=4.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlglot>=29.0.1",
    "sqlparse>=0.5.5",
]

[tool.uv]
# Ensure shared module is in the Python path
package = false

[tool.pytest.ini_options]
pythonpath = [".", "shared"]
